update ens transaction component and actions
